Prevent deletion of products associated with order items
- Added a check in the DELETE method of `product_detail` to prevent deletion of products that are linked to order items.
- Returns a 405 Method Not Allowed response with an error message if the product is associated with any order item.

+++ store/views.py
@@ -30,6 +30,8 @@ def product_detail(request, id):
         serializer.save()  
         return Response(serializer.data, status=status.HTTP_200_OK)  
     elif request.method == 'DELETE':
+        if product.orderitem_set.count() > 0:
+            return Response ({'error': 'This item cannot be deleted currently as it is associated with an order item'},status=status.HTTP_405_METHOD_NOT_ALLOWED)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
